requirements.txt

numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.1
tensorflow==2.13.0
torch==2.1.0
transformers==4.34.0
matplotlib==3.8.0
seaborn==0.13.0
opencv-python==4.8.1.78
jupyter==1.0.0
jupyterlab==4.0.5
flask==2.3.3
fastapi==0.95.2
uvicorn==0.23.2
requests==2.31.0
pytest==7.4.0
black==23.9.1
google-genai
pydantic
